added a section for user docs
the division of project for users and project for software developers is confusing anyone who finds the project because half of the project is always missing :( (the PIs dont want users to google for the app, and find the codebase because it will scare them). maybe adding links to the user docs will make this page more complete.